rename TimeTree::Client to TimeTree::OAuthApp::Client
