+++ lib/timetree/models/base_model.rb
@@ -15,7 +15,7 @@ module TimeTree
     # @param data [Hash]
     # TimeTree apis's response data.
     # @param included [Hash]
-    # @param client [TimeTree::Client]
+    # @param client [TimeTree::OAuthApp::Client]
     # @return [TimeTree::User, TimeTree::Label, TimeTree::Calendar, TimeTree::Event, TimeTree::Activity]
     # A TimeTree model object that be based on the type.
     # @raise [TimeTree::Error] if the type property is not set or unknown.
